Move DropSharedMemoryDeallocOp out of header (to squash)
